V0.2A
Fixed the single trigger on power up. via not forcing a start whle init the timer. Fixed DMA. Need to rewrite some of the menu to reflect the timing   structure. As of now there is not a need to fix this as I myself know how th product needs to be configured.