Enabled a more superior DMA mode for PWM timer Generatration
PWM timer geeration now used circular buffers to avoid the extra cycles needed to de-reinit the Normal DMA buffer after every transaction, this makes it easy to complete 1uS and faster transactions with little to no errors in timing. Overall VERY HAPPY with this, The next step is to incorperate a De-Init function and selection method to be able to witch between PWM and Triggered Modes. Also maybe Tiggered PWM? in the future?