Sushi Menu Overhaul and Optimization
Runing at abou 29KB out of 31KB feeling like a race to the bottom in terms of getting features added and such in such a tight constraint, The user of atof STDLIB function caused the usage of .rodata to spike by the order of 4KB so I made my own version of this function that is designed to handle positive numbers, with 3 decimal place precesison. The other note is saving this information gets weird as the HAL_Flash_Program seems to want a uint32t which defeats the purpose of having a float in the first place. These issues were resolved with some points.